Relocate AgentDiagnosisApi
