+++ server/web-service/src/main/java/org/bithon/server/web/service/diagnosis/AgentDiagnosisApi.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.web.service.agent.api;
+package org.bithon.server.web.service.diagnosis;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
