+++ server/web-service/src/main/java/org/bithon/server/web/service/diagnosis/JfrApi.java
@@ -36,7 +36,6 @@ import org.bithon.server.agent.controller.service.AgentControllerServer;
 import org.bithon.server.discovery.client.DiscoveredServiceInstance;
 import org.bithon.server.discovery.client.DiscoveredServiceInvoker;
 import org.bithon.server.discovery.declaration.controller.IAgentControllerApi;
-import org.bithon.server.web.service.agent.api.AgentDiagnosisApi;
 import org.bithon.server.web.service.agent.sql.table.AgentServiceProxyFactory;
 import org.bithon.shaded.com.google.protobuf.GeneratedMessageV3;
 import org.bithon.shaded.com.google.protobuf.util.JsonFormat;
