+++ server/web-service/src/test/java/org/bithon/server/web/service/diagnosis/AgentDiagnosisApiTest.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.web.service.agent.api;
+package org.bithon.server.web.service.diagnosis;
 
 
 import com.fasterxml.jackson.databind.ObjectMapper;
